% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bioc.R
\name{bioc}
\alias{bioc}
\title{A wrapper around the biocLite function, for the lazy}
\usage{
bioc(force_source = FALSE, ...)
}
\arguments{
\item{force_source}{\code{source} the biocLite.R script even if a biocLite
function is already available. Default: \code{FALSE}.}

\item{...}{Argument to pass to the biocLite function.}
}
\description{
This function will source the biocLite function from bioconductor website if
necessary and will then launch biocLite.
}
\details{
biocLite script is sourced from http://www.bioconductor.org/biocLite.R
}
\examples{
bioc() # To update all packages
bioc("metagene") # To install a Bioconductor's package
}

